+++ Constraid/LimitByMargin.swift
@@ -6,6 +6,126 @@ import UIKit
 extension ConstraidView {
     @discardableResult
     open func limit(byLeadingMarginOf item: Any?,
+        insetBy inset: CGFloat = 0.0,
+        multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        self.translatesAutoresizingMaskIntoConstraints = false
+        let collection = ConstraidConstraintCollection([
+            NSLayoutConstraint(item: self, attribute: .leading,
+                relatedBy: .lessThanOrEqual, toItem: item,
+                attribute: .leadingMargin, multiplier: multiplier,
+                constant: inset, priority: priority)
+            ])
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byTrailingMarginOf item: Any?,
+        insetBy inset: CGFloat = 0.0,
+        multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        self.translatesAutoresizingMaskIntoConstraints = false
+        let collection = ConstraidConstraintCollection([
+            NSLayoutConstraint(item: self, attribute: .trailing,
+               relatedBy: .lessThanOrEqual, toItem: item,
+               attribute: .trailingMargin, multiplier: multiplier,
+               constant: (inset * -1), priority: priority)
+            ])
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byTopMarginOf item: Any?,
+        insetBy inset: CGFloat = 0.0,
+        multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        self.translatesAutoresizingMaskIntoConstraints = false
+        let collection = ConstraidConstraintCollection([
+            NSLayoutConstraint(item: self, attribute: .top,
+                relatedBy: .lessThanOrEqual, toItem: item,
+                attribute: .topMargin, multiplier: multiplier,
+                constant: inset, priority: priority)
+            ])
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byBottomMarginOf item: Any?,
+        insetBy inset: CGFloat = 0.0, multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        self.translatesAutoresizingMaskIntoConstraints = false
+        let collection = ConstraidConstraintCollection([
+            NSLayoutConstraint(item: self, attribute: .bottom,
+                relatedBy: .lessThanOrEqual, toItem: item,
+                attribute: .bottomMargin, multiplier: multiplier,
+                constant: (inset * -1), priority: priority)
+            ])
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byHorizontalMarginsOf item: Any?,
+        insetBy inset: CGFloat = 0.0, multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        let collection = limit(byTopMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority) +
+                         limit(byBottomMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority)
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byVerticalMarginsOf item: Any?,
+        insetBy inset: CGFloat = 0.0, multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        let collection = limit(byLeadingMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority) +
+                         limit(byTrailingMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority)
+        collection.activate()
+        return collection
+    }
+
+    @discardableResult
+    open func limit(byMarginsOf item: Any?,
+        insetBy inset: CGFloat = 0.0, multiplier: CGFloat = 1.0,
+        priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
+        ) -> ConstraidConstraintCollection {
+
+        let collection = limit(byTopMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority) +
+                         limit(byBottomMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority) +
+                         limit(byLeadingMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority) +
+                         limit(byTrailingMarginOf: item, insetBy: inset,
+                             multiplier: multiplier, priority: priority)
+        collection.activate()
+        return collection
+    }
+
+    // MARK: - Deprecated
+
+    @discardableResult
+    @available(*, deprecated, message: "use limit(byLeadigMarginOf: , insetBy: ...)")
+    open func limit(byLeadingMarginOf item: Any?,
         constant: CGFloat = 0.0,
         multiplier: CGFloat = 1.0,
         priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
@@ -23,6 +143,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byTrailingMarginOf: , insetBy: ...)")
     open func limit(byTrailingMarginOf item: Any?,
         constant: CGFloat = 0.0,
         multiplier: CGFloat = 1.0,
@@ -41,6 +162,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byTopMarginOf: , insetBy: ...)")
     open func limit(byTopMarginOf item: Any?,
         constant: CGFloat = 0.0,
         multiplier: CGFloat = 1.0,
@@ -59,6 +181,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byBottomMarginOf: , insetBy: ...)")
     open func limit(byBottomMarginOf item: Any?,
         constant: CGFloat = 0.0, multiplier: CGFloat = 1.0,
         priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
@@ -76,6 +199,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byHorizontalMarginsOf: , insetBy: ...)")
     open func limit(byHorizontalMarginsOf item: Any?,
         constant: CGFloat = 0.0, multiplier: CGFloat = 1.0,
         priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
@@ -90,6 +214,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byVerticalMarginsOf: , insetBy: ...)")
     open func limit(byVerticalMarginsOf item: Any?,
         constant: CGFloat = 0.0, multiplier: CGFloat = 1.0,
         priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired
@@ -104,6 +229,7 @@ extension ConstraidView {
     }
 
     @discardableResult
+    @available(*, deprecated, message: "use limit(byMarginsOf: , insetBy: ...)")
     open func limit(byMarginsOf item: Any?,
         constant: CGFloat = 0.0, multiplier: CGFloat = 1.0,
         priority: ConstraidLayoutPriority = ConstraidLayoutPriorityRequired